Updated fast model with small feature changes
Former-commit-id: f4662536069d6b853bf237d8df87657827944a35 [formerly f9f3c606b3194a2d6827fa786680b0afb254ee8e] [formerly de6e97077b2d66f993e0cae1ffb78ea4b3cb9444 [formerly 679b617a2b093b7ce02a402fcba35d2354f7bb70]]
Former-commit-id: de6e97077b2d66f993e0cae1ffb78ea4b3cb9444
Former-commit-id: f5a696bbcfd1500dc64b7690a47b2b7c1fb50f17

+++ automaticityModel.m
@@ -684,6 +684,11 @@ end
 
 % Return set of parameters based on argument
 function [param_struct] = get_parameters(configuration)
+    % Necessary for codegen
+    % coder.extrinsic('cell2struct');
+    % Preinitialize param_struct to allow codegen to infer type
+    % param_struct = struct('PRE_LEARNING_TRIALS',0, 'LEARNING_TRIALS',0, 'POST_LEARNING_TRIALS',0, 'NOISE',0, 'PFC_DECISION_PT',0, 'PMC_DECISION_PT',0,'HEB_CONSTS',0,'ANTI_HEB_CONSTS',0,'NMDA',0,'AMPA',0,'W_MAX',0);
+    
     % Initialize parameters that depend on configuration
     param_names   = {'PRE_LEARNING_TRIALS'; 'LEARNING_TRIALS'; 'POST_LEARNING_TRIALS'; 'NOISE'; 'PFC_DECISION_PT'; 'PMC_DECISION_PT'};
     MADDOX_CONFIG = {                    0;               500;                      0;       0;                 4;                 4};
@@ -731,7 +736,7 @@ function [neuron_id, latency] = determine_reacting_neuron(n1, n2, decision_pt)
     % If latencies are equal (decision point never reached), take the
     % higher integral as the reacting neuron
     else
-        neuron_id = trapz(n1) < trapz(n2) + 1;
+        neuron_id = double(trapz(n1) < trapz(n2)) + 1;
         latency = length(n1);
     end
 end
